https://github.com/sffjunkie/astral/commit/ef270e7cae9deb3e6e5f31d9dd5d62e7aa85b18c
